End of project
They were modified: Modal, SecMov, Movies.
In order to be able to perform the delivery in time and form, to do
these three parts of the project, support me in the comits of my
colleagues, since I could not find a solution. All in order to learn and
understand how it worked.
With this commit I finish the last delivery of the Boot-camp.